commented and refactored database files and improved response formats

+++ api/database/comment.js
@@ -1,71 +1,65 @@
+/* This is controller module for comment data
+* This module performs CRUDE operation to database on comment table only
+*/
 module.exports = (connection) => {
 	const module = {};
 	module.createComment = async(req, res) => {
-		if(req.user) {
-			try {			
+		if (req.user) {
+			try {
 				const query = 'INSERT INTO comment (content, owner, parent_post) VALUES(?, ?, ?)';
 				const queryParams = [req.body.content, req.user.user_id, req.params.post_id];
-				const [rows, fields] = await connection.execute(query, queryParams);
-				res.send({message: 'Comnet created.'});
-			} catch(error) {
+				const [rows, _] = await connection.execute(query, queryParams);
+				res.send({message: 'Comment created.', comment: req.body.content, comment_id: rows.insertId});
+			} catch (error) {
 				res.status(401).json(error);
 			}
 		} else {
 			res.status(401).json({message: 'Unautherziez. Only loggedin user can comment.'});
 		}
-	}
+	};
 
 	module.getAllComments = async(req, res) => {
 		try {
-			const [rows, fields] = await connection.execute('SELECT * FROM comment WHERE parent_post=?',[req.params.post_id]);
+			const [rows, _] = await connection.execute('SELECT * FROM comment WHERE parent_post=?',[req.params.post_id]);
 			res.send(rows);
-		} catch(error) {
+		} catch (error) {
 			res.status(401).json(error);
 		}
-	}
+	};
 
 	module.updateComment = async(req, res) => {
-		if(req.user) {
+		if (req.user) {
 			try {
-				let query, queryParams, rows, fields;
-				if(req.user.admin_privileges) {
-					query = 'UPDATE comment SET content=? WHERE comment_id=?';
-					queryParams = [req.body.content, req.params.comment_id];
+				const query = req.user.admin_privileges ? 'UPDATE comment SET content=? WHERE comment_id=?' : 'UPDATE comment SET content=? WHERE comment_id=? AND owner=?';
+				const queryParams = req.user.admin_privileges ? [req.body.content, req.params.comment_id] : [req.body.content, req.params.comment_id, req.user.user_id];
+				const [rows, _] = await connection.execute(query, queryParams);
+				if (rows.affectedRows) {
+					res.send({message: 'Comment updated.', comment: req.body.content, comment_id: req.params.comment_id});
 				} else {
-					query = 'UPDATE comment SET content=? WHERE comment_id=? AND owner=?';
-					queryParams = [req.body.content, req.params.comment_id, req.user.user_id];
+					res.send({message: 'Comment does not exist or you don not have permission to delete'});
 				}
-				[rows, fields] = await connection.execute();
-				rows.affectedRows? res.send({message: 'Post delted.'}) : res.send({message: 'Post does not exist or you don not have permission to delete'});
-			} catch(error) {
+			} catch (error) {
 				res.status(401).json(error);
 			}
 		} else {
 			res.status(401).json('Unautherzied.');
 		}
-	}
+	};
 
 	module.deleteComment = async(req, res) => {
-		if(req.user) {
+		if (req.user) {
 			try {
-				let query, queryParams, rows, fields;
-				if(req.user.admin_privileges) {
-					query = 'DELETE FROM comment WHERE comment_id=?';
-					queryParams = [req.params.comment_id];
-				} else {
-					query = 'DELETE FROM comment WHERE comment_id=? AND owner=?';
-					queryParams = [req.params.comment_id, req.user.user_id];
-				}
-				[rows, fields] = await connection.execute();
-				rows.affectedRows? res.send({message: 'Post delted.'}) : res.send({message: 'Post does not exist or you don not have permission to delete'});
-				res.status(401).json(error);
-			} catch(error) {
+				const query = req.user.admin_privileges ? 'DELETE FROM comment WHERE comment_id=?' : 'DELETE FROM comment WHERE comment_id=? AND owner=?';
+				const queryParams = req.user.admin_privileges ? [req.params.comment_id] : [req.params.comment_id, req.user.user_id];
+				const [rows, _] = await connection.execute(query, queryParams);
+				rows.affectedRows ? res.send({message: 'Comment delted.'}) : res.send({message: 'Comment does not exist or you don not have permission to delete'});
+			} catch (error) {
 				res.send(error);
 			}
 		} else {
 			res.status(401).json('Unautherzied.');
 		}
-	}
+	};
 
 	return module;
-}
+};
